Remove operations in favor of commands only
