+++ src/main/java/com/nike/cerberus/command/Command.java
@@ -16,8 +16,6 @@
 
 package com.nike.cerberus.command;
 
-import com.nike.cerberus.operation.Operation;
-
 /**
  * Interface implemented by all commands available via the CLI.
  */
@@ -25,5 +23,7 @@ public interface Command {
 
     String getCommandName();
 
-    Class<? extends Operation<?>> getOperationClass();
+    void execute() throws Exception;
+
+    boolean isRunnable();
 }
