+++ src/main/java/com/nike/cerberus/command/StackDelegate.java
@@ -29,9 +29,6 @@ public class StackDelegate {
     public static final String OWNER_GROUP_LONG_ARG = "--owner-group";
     public static final String OWNER_EMAIL_LONG_ARG = "--owner-email";
     public static final String COST_CENTER_LONG_ARG = "--costcenter";
-    public static final String DESIRED_INSTANCES_LONG_ARG = "--desired-instances";
-    public static final String MAX_INSTANCES_LONG_ARG = "--max-instances";
-    public static final String MIN_INSTANCES_LONG_ARG = "--min-instances";
 
     @Parameter(names = AMI_ID_LONG_ARG, description = "The AMI ID for the specified stack.", required = true)
     private String amiId;
@@ -57,15 +54,6 @@ public class StackDelegate {
             required = true)
     private String costcenter;
 
-    @Parameter(names = DESIRED_INSTANCES_LONG_ARG, description = "Desired number of auto scaling instances.")
-    private int desiredInstances = 3;
-
-    @Parameter(names = MAX_INSTANCES_LONG_ARG, description = "Maximum number of auto scaling instances (must be larger than min).")
-    private int maximumInstances = 4;
-
-    @Parameter(names = MIN_INSTANCES_LONG_ARG, description = "Minimum number of auto scaling instances")
-    private int minimumInstances = 3;
-
     public String getAmiId() {
         return amiId;
     }
@@ -88,17 +76,5 @@ public class StackDelegate {
 
     public String getCostcenter() {
         return costcenter;
-    }
-
-    public int getDesiredInstances() {
-        return desiredInstances;
-    }
-
-    public int getMaximumInstances() {
-        return maximumInstances;
-    }
-
-    public int getMinimumInstances() {
-        return minimumInstances;
     }
 }
