+++ src/main/java/com/nike/cerberus/command/cms/CreateCmsConfigCommand.java
@@ -20,12 +20,16 @@ import com.beust.jcommander.DynamicParameter;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 import com.nike.cerberus.command.Command;
-import com.nike.cerberus.operation.Operation;
-import com.nike.cerberus.operation.cms.CreateCmsConfigOperation;
+import com.nike.cerberus.store.ConfigStore;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import javax.inject.Inject;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Properties;
 
+import static com.nike.cerberus.ConfigConstants.CMS_ADMIN_GROUP_KEY;
 import static com.nike.cerberus.command.cms.CreateCmsClusterCommand.COMMAND_NAME;
 
 /**
@@ -35,7 +39,9 @@ import static com.nike.cerberus.command.cms.CreateCmsClusterCommand.COMMAND_NAME
 public class CreateCmsConfigCommand implements Command {
 
     public static final String COMMAND_NAME = "create-cms-config";
+
     public static final String ADMIN_GROUP_LONG_ARG = "--admin-group";
+
     public static final String PROPERTY_SHORT_ARG = "-P";
 
     @Parameter(names = ADMIN_GROUP_LONG_ARG, description = "Group that has admin privileges in CMS.", required = true)
@@ -44,12 +50,36 @@ public class CreateCmsConfigCommand implements Command {
     @DynamicParameter(names = PROPERTY_SHORT_ARG, description = "Dynamic parameters for setting additional properties in the CMS environment configuration.")
     private Map<String, String> additionalProperties = new HashMap<>();
 
-    public String getAdminGroup() {
-        return adminGroup;
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
+    private final ConfigStore configStore;
+
+    @Inject
+    public CreateCmsConfigCommand(final ConfigStore configStore) {
+        this.configStore = configStore;
     }
 
-    public Map<String, String> getAdditionalProperties() {
-        return additionalProperties;
+    @Override
+    public void execute() {
+        configStore.storeCmsAdminGroup(adminGroup);
+
+        logger.info("Retrieving configuration data from the configuration bucket.");
+        final Properties cmsConfigProperties = configStore.getCmsSystemProperties();
+
+        cmsConfigProperties.put(CMS_ADMIN_GROUP_KEY, adminGroup);
+
+        additionalProperties.forEach((k, v) -> {
+            if (!cmsConfigProperties.containsKey(k)) {
+                cmsConfigProperties.put(k, v);
+            } else {
+                logger.warn("Ignoring additional property that would override system configured property, " + k);
+            }
+        });
+
+        logger.info("Uploading the CMS configuration to the configuration bucket.");
+        configStore.storeCmsEnvConfig(cmsConfigProperties);
+
+        logger.info("Uploading complete.");
     }
 
     @Override
@@ -58,7 +88,12 @@ public class CreateCmsConfigCommand implements Command {
     }
 
     @Override
-    public Class<? extends Operation<?>> getOperationClass() {
-        return CreateCmsConfigOperation.class;
-    }
-}
+    public boolean isRunnable() {
+        boolean isRunnable = !configStore.getCmsEnvConfig().isPresent();
+
+        if (! isRunnable) {
+            logger.warn("CMS config already exists, use 'update-cms-config' command.");
+        }
+
+        return isRunnable;
+    }}
