+++ src/main/java/com/nike/cerberus/command/core/UpdateStackCommand.java
@@ -16,22 +16,48 @@
 
 package com.nike.cerberus.command.core;
 
+import com.amazonaws.AmazonServiceException;
+import com.amazonaws.services.cloudformation.model.StackStatus;
 import com.beust.jcommander.DynamicParameter;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Sets;
+import com.nike.cerberus.ConfigConstants;
 import com.nike.cerberus.command.Command;
 import com.nike.cerberus.command.StackDelegate;
+import com.nike.cerberus.domain.cloudformation.BaseParameters;
+import com.nike.cerberus.domain.cloudformation.CmsParameters;
+import com.nike.cerberus.domain.cloudformation.LaunchConfigParameters;
+import com.nike.cerberus.domain.cloudformation.SslConfigParametersDelegate;
+import com.nike.cerberus.domain.cloudformation.TagParameters;
 import com.nike.cerberus.domain.environment.StackName;
-import com.nike.cerberus.operation.Operation;
-import com.nike.cerberus.operation.core.UpdateStackOperation;
+import com.nike.cerberus.command.UnexpectedCloudFormationStatusException;
+import com.nike.cerberus.service.AmiTagCheckService;
+import com.nike.cerberus.service.CloudFormationService;
+import com.nike.cerberus.service.Ec2UserDataService;
+import com.nike.cerberus.store.ConfigStore;
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import javax.inject.Inject;
+import javax.inject.Named;
 import java.util.HashMap;
 import java.util.Map;
 
+import static com.amazonaws.services.cloudformation.model.StackStatus.UPDATE_COMPLETE;
+import static com.amazonaws.services.cloudformation.model.StackStatus.UPDATE_COMPLETE_CLEANUP_IN_PROGRESS;
+import static com.amazonaws.services.cloudformation.model.StackStatus.UPDATE_ROLLBACK_COMPLETE;
+import static com.amazonaws.services.cloudformation.model.StackStatus.UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS;
+import static com.amazonaws.services.cloudformation.model.StackStatus.UPDATE_ROLLBACK_FAILED;
+import static com.nike.cerberus.ConfigConstants.CERT_PART_PUBKEY;
 import static com.nike.cerberus.command.core.UpdateStackCommand.COMMAND_NAME;
 import static com.nike.cerberus.ConfigConstants.SKIP_AMI_TAG_CHECK_ARG;
 import static com.nike.cerberus.ConfigConstants.SKIP_AMI_TAG_CHECK_DESCRIPTION;
-
+import static com.nike.cerberus.module.CerberusModule.CF_OBJECT_MAPPER;
 
 /**
  * Command for updating the specified CloudFormation stack with the new parameters.
@@ -39,8 +65,12 @@ import static com.nike.cerberus.ConfigConstants.SKIP_AMI_TAG_CHECK_DESCRIPTION;
 @Parameters(commandNames = COMMAND_NAME, commandDescription = "Updates the specified CloudFormation stack.")
 public class UpdateStackCommand implements Command {
 
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
     public static final String COMMAND_NAME = "update-stack";
+
     public static final String OVERWRITE_TEMPLATE_LONG_ARG = "--overwrite-template";
+
     public static final String PARAMETER_SHORT_ARG = "-P";
 
     @Parameter(names = {"--stack-name"}, required = true, description = "The stack name to update.")
@@ -72,15 +102,6 @@ public class UpdateStackCommand implements Command {
             description = "Flag for overwriting existing CloudFormation template")
     private boolean overwriteTemplate;
 
-    @Parameter(names = StackDelegate.DESIRED_INSTANCES_LONG_ARG, description = "Desired number of auto scaling instances.")
-    private Integer desiredInstances;
-
-    @Parameter(names = StackDelegate.MAX_INSTANCES_LONG_ARG, description = "Maximum number of auto scaling instances.")
-    private Integer maximumInstances;
-
-    @Parameter(names = StackDelegate.MIN_INSTANCES_LONG_ARG, description = "Minimum number of autos scaling instances")
-    private Integer minimumInstances;
-
     @Parameter(names = SKIP_AMI_TAG_CHECK_ARG,
             description = SKIP_AMI_TAG_CHECK_DESCRIPTION)
     private boolean skipAmiTagCheck;
@@ -88,56 +109,113 @@ public class UpdateStackCommand implements Command {
     @DynamicParameter(names = PARAMETER_SHORT_ARG, description = "Dynamic parameters for overriding the values for specific parameters in the CloudFormation.")
     private Map<String, String> dynamicParameters = new HashMap<>();
 
-    public StackName getStackName() {
-        return stackName;
+    private final Map<StackName, Class<? extends LaunchConfigParameters>> stackParameterMap;
+
+    private final Map<StackName, String> stackTemplatePathMap;
+
+    private final ConfigStore configStore;
+
+    private final CloudFormationService cloudFormationService;
+
+    private final ObjectMapper cloudFormationObjectMapper;
+
+    private final Ec2UserDataService ec2UserDataService;
+
+    private final AmiTagCheckService amiTagCheckService;
+
+    @Inject
+    public UpdateStackCommand(final ConfigStore configStore,
+                                final CloudFormationService cloudFormationService,
+                                @Named(CF_OBJECT_MAPPER) final ObjectMapper cloudFormationObjectMapper,
+                                final Ec2UserDataService ec2UserDataService,
+                                final AmiTagCheckService amiTagCheckService) {
+        this.configStore = configStore;
+        this.cloudFormationService = cloudFormationService;
+        this.cloudFormationObjectMapper = cloudFormationObjectMapper;
+        this.ec2UserDataService = ec2UserDataService;
+        this.amiTagCheckService = amiTagCheckService;
+
+        stackParameterMap = new HashMap<>();
+        stackParameterMap.put(StackName.CMS, CmsParameters.class);
+
+        stackTemplatePathMap = new HashMap<>();
+        stackTemplatePathMap.put(StackName.BASE, ConfigConstants.VPC_STACK_TEMPLATE_PATH);
+        stackTemplatePathMap.put(StackName.CMS, ConfigConstants.CMS_STACK_TEMPLATE_PATH);
     }
 
-    public String getOwnerGroup() {
-        return ownerGroup;
+    @Override
+    public void execute() {
+        final String stackId = configStore.getStackId(stackName);
+        final Class<? extends LaunchConfigParameters> parametersClass = stackParameterMap.get(stackName);
+        final Map<String, String> parameters;
+
+        if (parametersClass != null) {
+            parameters = getUpdateLaunchConfigParameters(stackName, parametersClass);
+        } else if (StackName.BASE == stackName) {
+            parameters = getUpdatedBaseStackParameters();
+        } else {
+            throw new IllegalArgumentException("The specified stack does not support the update stack command!");
+        }
+
+        // Make sure the given AmiId is for this component. Check if it contains required tag
+        // There is no AMI for Base.
+        if ( ! skipAmiTagCheck && StackName.BASE != stackName ) {
+            amiTagCheckService.validateAmiTagForStack(amiId,stackName);
+        }
+
+        parameters.putAll(dynamicParameters);
+
+        try {
+            logger.info("Starting the update for {}.", stackName.getName());
+
+            if (overwriteTemplate) {
+                cloudFormationService.updateStack(stackId, parameters, stackTemplatePathMap.get(stackName), true);
+            } else {
+                cloudFormationService.updateStack(stackId, parameters, true);
+            }
+
+            final StackStatus endStatus =
+                    cloudFormationService.waitForStatus(stackId,
+                            Sets.newHashSet(
+                                    UPDATE_COMPLETE,
+                                    UPDATE_COMPLETE_CLEANUP_IN_PROGRESS,
+                                    UPDATE_ROLLBACK_COMPLETE,
+                                    UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS,
+                                    UPDATE_ROLLBACK_FAILED
+                            ));
+
+            if (! ImmutableList.of(UPDATE_COMPLETE, UPDATE_COMPLETE_CLEANUP_IN_PROGRESS).contains(endStatus)) {
+                final String errorMessage = String.format("CloudFormation reports that updating the stack was not successful. end status: %s", endStatus.name());
+                logger.error(errorMessage);
+
+                throw new UnexpectedCloudFormationStatusException(errorMessage);
+            }
+
+            logger.info("Update complete.");
+        } catch (AmazonServiceException ase) {
+            if (ase.getStatusCode() == 400 &&
+                    StringUtils.equalsIgnoreCase(ase.getErrorMessage(), "No updates are to be performed.")) {
+                logger.warn("CloudFormation reported no changes detected.");
+            } else {
+                throw ase;
+            }
+        }
     }
 
-    public String getAmiId() {
-        return amiId;
-    }
+    @Override
+    public boolean isRunnable() {
+        boolean isRunnable = true;
+        final String stackId = configStore.getStackId(stackName);
 
-    public String getInstanceSize() {
-        return instanceSize;
-    }
+        if (StringUtils.isBlank(stackId)) {
+            logger.error("The stack name specified has not been created for this environment yet!");
+            isRunnable = false;
+        } else if (!cloudFormationService.isStackPresent(stackId)) {
+            logger.error("CloudFormation doesn't have the specified stack: {}", stackId);
+            isRunnable = false;
+        }
 
-    public String getKeyPairName() {
-        return keyPairName;
-    }
-
-    public String getOwnerEmail() {
-        return ownerEmail;
-    }
-
-    public String getCostcenter() {
-        return costcenter;
-    }
-
-    public boolean isOverwriteTemplate() {
-        return overwriteTemplate;
-    }
-
-    public Map<String, String> getDynamicParameters() {
-        return dynamicParameters;
-    }
-
-    public Integer getDesiredInstances() {
-        return desiredInstances;
-    }
-
-    public Integer getMaximumInstances() {
-        return maximumInstances;
-    }
-
-    public Integer getMinimumInstances() {
-        return minimumInstances;
-    }
-
-    public boolean isSkipAmiTagCheck() {
-        return skipAmiTagCheck;
+        return isRunnable;
     }
 
     @Override
@@ -145,8 +223,70 @@ public class UpdateStackCommand implements Command {
         return COMMAND_NAME;
     }
 
-    @Override
-    public Class<? extends Operation<?>> getOperationClass() {
-        return UpdateStackOperation.class;
+    private Map<String, String> getUpdateLaunchConfigParameters(final StackName stackName,
+                                                                final Class<? extends LaunchConfigParameters> parametersClass) {
+        final LaunchConfigParameters launchConfigParameters =
+                configStore.getStackParameters(stackName, parametersClass);
+
+        launchConfigParameters.getLaunchConfigParameters().setUserData(
+                ec2UserDataService.getUserData(stackName, ownerGroup));
+
+        if (StringUtils.isNotBlank(amiId)) {
+            launchConfigParameters.getLaunchConfigParameters().setAmiId(amiId);
+        }
+
+        if (StringUtils.isNotBlank(instanceSize)) {
+            launchConfigParameters.getLaunchConfigParameters().setInstanceSize(instanceSize);
+        }
+
+        if (StringUtils.isNotBlank(keyPairName)) {
+            launchConfigParameters.getLaunchConfigParameters().setKeyPairName(keyPairName);
+        }
+
+        if (StringUtils.isNotBlank(ownerEmail)) {
+            launchConfigParameters.getTagParameters().setTagEmail(ownerEmail);
+        }
+
+        if (StringUtils.isNotBlank(costcenter)) {
+            launchConfigParameters.getTagParameters().setTagCostcenter(costcenter);
+        }
+
+        updateSslConfigParameters(stackName, launchConfigParameters);
+
+        final TypeReference<Map<String, String>> typeReference = new TypeReference<Map<String, String>>() {};
+        return cloudFormationObjectMapper.convertValue(launchConfigParameters, typeReference);
+    }
+
+    private void updateSslConfigParameters(final StackName stackName,
+                                           final LaunchConfigParameters launchConfigParameters) {
+
+        final SslConfigParametersDelegate sslConfigParameters = launchConfigParameters.getSslConfigParameters();
+
+        if (StringUtils.isNotBlank(sslConfigParameters.getCertPublicKey())) {
+            sslConfigParameters.setCertPublicKey(configStore.getCertPart(stackName, CERT_PART_PUBKEY).get());
+        }
+
+        if (StringUtils.isNotBlank(sslConfigParameters.getSslCertificateId())) {
+            sslConfigParameters.setSslCertificateId(configStore.getServerCertificateId(stackName).get());
+        }
+
+        if (StringUtils.isNotBlank(sslConfigParameters.getSslCertificateArn())) {
+            sslConfigParameters.setSslCertificateArn(configStore.getServerCertificateArn(stackName).get());
+        }
+    }
+
+    private Map<String, String> getUpdatedBaseStackParameters() {
+        final TagParameters tagParameters = configStore.getStackParameters(stackName, BaseParameters.class);
+
+        if (StringUtils.isNotBlank(ownerEmail)) {
+            tagParameters.getTagParameters().setTagEmail(ownerEmail);
+        }
+
+        if (StringUtils.isNotBlank(costcenter)) {
+            tagParameters.getTagParameters().setTagCostcenter(costcenter);
+        }
+
+        final TypeReference<Map<String, String>> typeReference = new TypeReference<Map<String, String>>() {};
+        return cloudFormationObjectMapper.convertValue(tagParameters, typeReference);
     }
 }
