+++ src/main/java/com/nike/cerberus/command/core/ViewConfigCommand.java
@@ -19,8 +19,14 @@ package com.nike.cerberus.command.core;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 import com.nike.cerberus.command.Command;
-import com.nike.cerberus.operation.Operation;
-import com.nike.cerberus.operation.core.ViewConfigOperation;
+import com.nike.cerberus.store.ConfigStore;
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.inject.Inject;
+import java.util.Optional;
+import java.util.Set;
 
 import static com.nike.cerberus.command.cms.CreateCmsClusterCommand.COMMAND_NAME;
 
@@ -30,22 +36,50 @@ import static com.nike.cerberus.command.cms.CreateCmsClusterCommand.COMMAND_NAME
 @Parameters(commandNames = COMMAND_NAME, commandDescription = "Shows configuration files from S3.")
 public class ViewConfigCommand implements Command {
 
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
     public static final String COMMAND_NAME = "view-config";
 
     @Parameter(names = {"--config-path"}, required = true, description = "The path to the configuration file (e.g. data/cms/environment.properties)")
     private String pathToConfig;
 
-    public String getPathToConfig() {
-        return pathToConfig;
+    private final ConfigStore configStore;
+
+    @Inject
+    public ViewConfigCommand(final ConfigStore configStore) {
+        this.configStore = configStore;
+    }
+
+    @Override
+    public void execute() {
+
+        String path = pathToConfig;
+        if (StringUtils.startsWith(path, "/")) {
+            logger.warn("Path started with leading slash, removing...");
+            path = StringUtils.stripStart(path, "/");
+        }
+
+        Optional<String> fileContents = configStore.getConfigProperties(path);
+        if (fileContents.isPresent()) {
+            logger.info(fileContents.get());
+        } else {
+            Set<String> keys = configStore.listUnderPartialPath(path);
+            if (!keys.isEmpty()) {
+                logger.info("List under path '{}': {}", path, keys);
+            }
+            else {
+                logger.error(String.format("Failed to load file: '%s'", path));
+            }
+        }
+    }
+
+    @Override
+    public boolean isRunnable() {
+        return true;
     }
 
     @Override
     public String getCommandName() {
         return COMMAND_NAME;
-    }
-
-    @Override
-    public Class<? extends Operation<?>> getOperationClass() {
-        return ViewConfigOperation.class;
     }
 }
