+++ src/main/java/com/nike/cerberus/domain/cloudformation/VpcOutputs.java
@@ -1,0 +1,111 @@
+/*
+ * Copyright (c) 2016 Nike, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.nike.cerberus.domain.cloudformation;
+
+/**
+ * Represents the base stack outputs.
+ */
+public class VpcOutputs {
+
+    private String vpcId;
+
+    private String vpcHostedZoneId;
+
+    private String vpcSubnetIdForAz1;
+
+    private String vpcSubnetIdForAz2;
+
+    private String vpcSubnetIdForAz3;
+
+    private String subnetCidrBlockForAz1;
+
+    private String subnetCidrBlockForAz2;
+
+    private String subnetCidrBlockForAz3;
+
+    public String getVpcId() {
+        return vpcId;
+    }
+
+    public VpcOutputs setVpcId(String vpcId) {
+        this.vpcId = vpcId;
+        return this;
+    }
+
+    public String getVpcHostedZoneId() {
+        return vpcHostedZoneId;
+    }
+
+    public VpcOutputs setVpcHostedZoneId(String vpcHostedZoneId) {
+        this.vpcHostedZoneId = vpcHostedZoneId;
+        return this;
+    }
+
+    public String getVpcSubnetIdForAz1() {
+        return vpcSubnetIdForAz1;
+    }
+
+    public VpcOutputs setVpcSubnetIdForAz1(String vpcSubnetIdForAz1) {
+        this.vpcSubnetIdForAz1 = vpcSubnetIdForAz1;
+        return this;
+    }
+
+    public String getVpcSubnetIdForAz2() {
+        return vpcSubnetIdForAz2;
+    }
+
+    public VpcOutputs setVpcSubnetIdForAz2(String vpcSubnetIdForAz2) {
+        this.vpcSubnetIdForAz2 = vpcSubnetIdForAz2;
+        return this;
+    }
+
+    public String getVpcSubnetIdForAz3() {
+        return vpcSubnetIdForAz3;
+    }
+
+    public VpcOutputs setVpcSubnetIdForAz3(String vpcSubnetIdForAz3) {
+        this.vpcSubnetIdForAz3 = vpcSubnetIdForAz3;
+        return this;
+    }
+
+    public String getSubnetCidrBlockForAz1() {
+        return subnetCidrBlockForAz1;
+    }
+
+    public VpcOutputs setSubnetCidrBlockForAz1(String subnetCidrBlockForAz1) {
+        this.subnetCidrBlockForAz1 = subnetCidrBlockForAz1;
+        return this;
+    }
+
+    public String getSubnetCidrBlockForAz2() {
+        return subnetCidrBlockForAz2;
+    }
+
+    public VpcOutputs setSubnetCidrBlockForAz2(String subnetCidrBlockForAz2) {
+        this.subnetCidrBlockForAz2 = subnetCidrBlockForAz2;
+        return this;
+    }
+
+    public String getSubnetCidrBlockForAz3() {
+        return subnetCidrBlockForAz3;
+    }
+
+    public VpcOutputs setSubnetCidrBlockForAz3(String subnetCidrBlockForAz3) {
+        this.subnetCidrBlockForAz3 = subnetCidrBlockForAz3;
+        return this;
+    }
+}
