+++ src/main/java/com/nike/cerberus/store/ConfigStore.java
@@ -277,31 +277,6 @@ public class ConfigStore {
     }
 
     /**
-     * Retrieves the CMS Vault token from the config store.
-     *
-     * @return CMS Vault token
-     */
-    public Optional<String> getCmsVaultToken() {
-        synchronized (secretsDataLock) {
-            final Secrets secrets = getSecretsData();
-            return Optional.ofNullable(secrets.getCms().getVaultToken());
-        }
-    }
-
-    /**
-     * Stores the CMS Vault token.
-     *
-     * @param cmsVaultToken CMS Vault token
-     */
-    public void storeCmsVaultToken(final String cmsVaultToken) {
-        synchronized (secretsDataLock) {
-            final Secrets secrets = getSecretsData();
-            secrets.getCms().setVaultToken(cmsVaultToken);
-            saveSecretsData(secrets);
-        }
-    }
-
-    /**
      * Gets the Vault root token from the config store.
      *
      * @return Vault root token
@@ -398,7 +373,7 @@ public class ConfigStore {
     private Properties generateBaseCmsSystemProperties() {
 
         final BaseOutputs baseOutputs = getBaseStackOutputs();
-        final BaseParameters baseParameters = getBaseStackParameters();
+//        final BaseParameters baseParameters = getBaseStackParameters();
         final Optional<String> cmsDatabasePassword = getCmsDatabasePassword();
 
         final GetCallerIdentityResult callerIdentity = securityTokenService.getCallerIdentity(
@@ -407,7 +382,7 @@ public class ConfigStore {
 
         final Properties properties = new Properties();
         properties.put(ROOT_USER_ARN_KEY, rootUserArn);
-        properties.put(ADMIN_ROLE_ARN_KEY, baseParameters.getAccountAdminArn());
+//        properties.put(ADMIN_ROLE_ARN_KEY, baseParameters.getAccountAdminArn());
         properties.put(CMS_ROLE_ARN_KEY, baseOutputs.getCmsIamRoleArn());
         properties.put(JDBC_URL_KEY, baseOutputs.getCmsDbJdbcConnectionString());
         properties.put(JDBC_USERNAME_KEY, ConfigConstants.DEFAULT_CMS_DB_NAME);
@@ -417,8 +392,7 @@ public class ConfigStore {
     }
 
     public Optional<String> getAccountAdminArn() {
-        final BaseParameters baseParameters = getBaseStackParameters();
-        return Optional.ofNullable(baseParameters.getAccountAdminArn());
+        throw new UnsupportedOperationException();
     }
 
     public String getCerberusBaseUrl() {
@@ -576,19 +550,6 @@ public class ConfigStore {
         }
     }
 
-    /**
-     * Constructs the standard format for CNAMEs used by internal ELBs.
-     *
-     * @param stackName Stack name the CNAME is for
-     * @return CNAME
-     */
-    public String getInternalElbCname(final StackName stackName) {
-        final BaseParameters baseParameters = getBaseStackParameters();
-        return String.format("%s.%s.%s.",
-                stackName.getName(),
-                environmentMetadata.getRegionName(),
-                baseParameters.getVpcHostedZoneName());
-    }
 
     /**
      * Initializes the environment data in the config bucket.
