+++ src/main/resources/cloudformation/vpc.yaml
@@ -22,13 +22,6 @@ Outputs:
     Description: The VPC subnet in AZ '3'
     Value: !Ref 'CerberusSubnetForAz3'
 Parameters:
-  internetGatewayCidrBlock:
-    AllowedPattern: '[0-9./]*'
-    Default: 0.0.0.0/0
-    Description: The application load balancer CIDR block for where traffic is allowed from
-    MaxLength: '20'
-    MinLength: '9'
-    Type: String
   az1:
     AllowedPattern: '[a-z]{2}-[a-z]+-\d\w'
     Description: An availability zone identifier for zone '1'
@@ -40,6 +33,13 @@ Parameters:
   az3:
     AllowedPattern: '[a-z]{2}-[a-z]+-\d\w'
     Description: An availability zone identifier for zone '3'
+    Type: String
+  internetGatewayCidrBlock:
+    AllowedPattern: '[0-9./]*'
+    Default: 0.0.0.0/0
+    Description: The application load balancer CIDR block for where traffic is allowed from
+    MaxLength: '20'
+    MinLength: '9'
     Type: String
   subnetCidrBlockForAz1:
     AllowedPattern: '[0-9./]*'
